Crate pangalacticcc

A helpful calculator for hitchhiking merchants all across the galaxy and beyond. Converts alien numerals and units to Credits. Input is read from stdin or text file. Output is printed to stdout.

Modules

roman

textprocessing

Structs

Config

Holds the path of the file to be processed as String.

Functions

answer_how_many_credits

Returns response to input asking “how many credits is …” as String

answer_how_much

Returns response to input asking “how much is …” as String

get_args

Parses command line arguments

open

Returns a BufReader for path on success. If path is "-", BufReader for stdin is returned, otherwise BufReader for file path is returned.

run

Runs the program on provided config. Output is printed to stdout