Module pangalacticcc::textprocessing

Functions

extract_amount_credits_from_sentence

Returns a amount of Credits extracted from a sentence or None if extraction failed.

extract_amounts_from_sentence

Returns a amount extracted from a sentence with numerals as integer or None if extraction failed. Sentences are expected to also have an amount stated directly before the unit.

extract_unit_values_from_sentence

Returns credit conversion rate for unit extracted from a sentence with numerals as (unit: String, value: f64) tuple or None if extraction failed. Sentences are expected to have an amount stated directly before the unit.

extract_units_from_sentence

Returns a unit extracted from a sentence about unit info or None if extraction failed. Sentences are expected to also have an amount stated directly before the unit.

is_numeral_info

Returns true if sentence is statement about numerals

is_question_how_many_credits

Returns true if questions asks for how many credits, and an alien numeral can be extracted

is_question_how_much

Returns true if questions asks how much is

is_unit_info

Returns true if sentence is statement about units, and a number of Credits can be extracted

numerals_to_roman

Returns (alien_numeral, roman_numeral) tuple from a sentence with numerals or None if extraction failed.